Patch 1.2.1 part 4
The second to last part of 1.2.1. There are now 10 different stages to choose from and the colliding with walls code has been reworked to that if you're not in stun but moving at a fast speed, you bounce off the wall.

The 10 stages need to be made and the stage selection screen needs to be reworked